refactor: move js files into /src
