+++ gameSite.html
@@ -7,7 +7,7 @@
 
   <body>
     <!-- <script type="module" src="./helperFunctions.js"></script> -->
-    <script type="module" src="./gameIncantation.js"></script>
+    <script type="module" src="./src/gameIncantation.js"></script>
     <!-- <script src="./summoningUI.js"></script> -->
   </body>
 </html>
