+++ package.json
@@ -23,6 +23,7 @@
     "eslint": "^7.6.0",
     "eslint-config-airbnb-base": "^14.2.0",
     "eslint-plugin-import": "^2.22.0",
+    "jest": "^27.4.5",
     "parcel": "^2.0.1"
   }
 }
